Splitting the monetizable specs into separate files for better organization
